Fix numeric data types

+++ WeatherApp.Core/Models/WeatherDetails.cs
@@ -18,22 +18,22 @@
         /// <summary>
         /// Temperature.
         /// </summary>
-        public float Temperature { get; set; }
+        public double Temperature { get; set; }
 
         /// <summary>
         /// Temperature according to the human perception of weather.
         /// </summary>
-        public float TemperatureFeelsLike { get; set; }
+        public double TemperatureFeelsLike { get; set; }
 
         /// <summary>
         /// Minimum temperature at the moment.
         /// </summary>
-        public float TemperatureMin { get; set; }
+        public double TemperatureMin { get; set; }
 
         /// <summary>
         /// Maximum temperature at the moment.
         /// </summary>
-        public float TemperatureMax { get; set; }
+        public double TemperatureMax { get; set; }
 
     }
 }
